add implementation of ViewClient to GUI
