+++ src/main/java/seedu/address/logic/Logic.java
@@ -33,6 +33,8 @@ public interface Logic {
     /** Returns an unmodifiable view of the filtered list of clients */
     ObservableList<Client> getFilteredClientList();
 
+    Client getClientInView();
+    
     /**
      * Returns the user prefs' address book file path.
      */
