+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -62,6 +62,11 @@ public class LogicManager implements Logic {
     }
 
     @Override
+    public Client getClientInView() {
+        return model.getClientInView();
+    }
+
+    @Override
     public Path getAddressBookFilePath() {
         return model.getAddressBookFilePath();
     }
