+++ src/main/java/seedu/address/ui/ViewClient.java
@@ -1,0 +1,131 @@
+package seedu.address.ui;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+import javafx.scene.layout.FlowPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.Region;
+import seedu.address.model.client.Client;
+
+/**
+ * An UI component that displays detailed information of a {@code Client}.
+ */
+public class ViewClient extends UiPart<Region> {
+
+    private static final String FXML = "ViewClientCard.fxml";
+
+    /**
+     * Note: Certain keywords such as "location" and "resources" are reserved
+     * keywords in JavaFX. As a consequence, UI elements' variable names cannot be
+     * set to such keywords or an exception will be thrown by JavaFX during runtime.
+     *
+     * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The
+     *      issue on AddressBook level 4</a>
+     */
+
+    public Client client;
+
+    @FXML
+    private HBox cardPane;
+    @FXML
+    private Label name;
+    @FXML
+    private Label phone;
+    @FXML
+    private Label address;
+    @FXML
+    private Label email;
+    @FXML
+    private Label gender;
+    @FXML
+    private Label birthday;
+    @FXML
+    private Label currentWeight;
+    @FXML
+    private Label targetWeight;
+    @FXML
+    private Label height;
+    @FXML
+    private Label remark;
+    @FXML
+    private FlowPane sports;
+    @FXML
+    private FlowPane tags;
+
+    public ViewClient(Client client) {
+        super(FXML);
+        if(client == null){
+            name.setText("");
+            phone.setText("");
+            address.setText("");
+            email.setText("");
+            gender.setText("");
+            birthday.setText("");
+            height.setText("");
+            currentWeight.setText("");
+            targetWeight.setText("");
+            remark.setText("");
+        } else {
+            this.client = client;
+            name.setText(client.getName().fullName);
+            String phoneLabel = "Phone: ";
+            String phoneAttributeForDisplay = getAttributeForDisplay(client.getPhone().value);
+            String fullPhoneForDisplay = phoneLabel + phoneAttributeForDisplay;
+            phone.setText(fullPhoneForDisplay);
+            String addressLabel = "Address: ";
+            String addressAttributeForDisplay = getAttributeForDisplay(client.getAddress().value);
+            String fullAddressForDisplay = addressLabel + addressAttributeForDisplay;
+            address.setText(fullAddressForDisplay);
+            String emailLabel = "Email: ";
+            String emailAttributeForDisplay = getAttributeForDisplay(client.getEmail().value);
+            String fullEmailForDisplay = emailLabel + emailAttributeForDisplay;
+            email.setText(fullEmailForDisplay);
+            String genderLabel = "Gender: ";
+            String genderAttributeForDisplay = getAttributeForDisplay(client.getGender().value);
+            String fullGenderForDisplay = genderLabel + genderAttributeForDisplay;
+            gender.setText(fullGenderForDisplay);
+            String birthdayLabel = "Birthday: ";
+            String birthdayAttributeForDisplay = getAttributeForDisplay(client.getBirthdayString());
+            String fullBirthdayForDisplay = birthdayLabel + birthdayAttributeForDisplay;
+            birthday.setText(fullBirthdayForDisplay);
+            String heightLabel = "Height: ";
+            String heightAttributeForDisplay = getAttributeForDisplay(client.getHeight().value);
+            String fullHeightForDisplay = heightLabel + heightAttributeForDisplay;
+            height.setText(fullHeightForDisplay);
+            String currentWeightLabel = "Current Weight: ";
+            String currentWeightAttributeForDisplay = getAttributeForDisplay(client.getCurrentWeight().value);
+            String fullCurrentWeightForDisplay = currentWeightLabel + currentWeightAttributeForDisplay;
+            currentWeight.setText(fullCurrentWeightForDisplay);
+            String targetWeightLabel = "Target Weight: ";
+            String targetWeightAttributeForDisplay = getAttributeForDisplay(client.getTargetWeight().value);
+            String fullTargetWeightForDisplay = targetWeightLabel + targetWeightAttributeForDisplay;
+            targetWeight.setText(fullTargetWeightForDisplay);
+            String remarkLabel = "Remark: ";
+            String remarkAttributeForDisplay = getAttributeForDisplay(client.getRemark().value);
+            String fullRemarkForDisplay = remarkLabel + remarkAttributeForDisplay;
+            remark.setText(fullRemarkForDisplay);
+        }
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        // short circuit if same object
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof ViewClient)) {
+            return false;
+        }
+
+        // state check
+        ViewClient card = (ViewClient) other;
+        return client.equals(card.client);
+    }
+
+    private String getAttributeForDisplay(String string) {
+        String emptyString = "";
+        return !string.equals(emptyString) ? string : "-";
+    }
+}
