+++ src/main/resources/view/MainWindow.fxml
@@ -12,9 +12,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.stage.Stage?>
 
-<fx:root minHeight="600" minWidth="1050.0" onCloseRequest="#handleExit" title="FitBiz" type="javafx.stage.Stage"
-  xmlns="http://javafx.com/javafx/11.0.1"
-  xmlns:fx="http://javafx.com/fxml/1">
+<fx:root minHeight="600" minWidth="1050.0" onCloseRequest="#handleExit" title="FitBiz" type="javafx.stage.Stage" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
   <icons>
     <Image url="@/images/bicep.png" />
   </icons>
@@ -57,14 +55,12 @@
                 <StackPane fx:id="clientListPanelPlaceholder" prefHeight="433.0" prefWidth="335.0" VBox.vgrow="ALWAYS" />
               </children>
             </VBox>
-            <VBox fx:id="viewClient" layoutX="10.0" layoutY="10.0" minWidth="450.0" prefHeight="455.0" prefWidth="400.0" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-              <padding>
-                <Insets bottom="10" left="10" right="10" top="10" />
-              </padding>
-              <children>
-                <StackPane fx:id="viewClientPanelPlaceholder" prefHeight="433.0" prefWidth="276.0" VBox.vgrow="ALWAYS" />
-              </children>
-            </VBox>
+                              <StackPane fx:id="viewClientPanelPlaceholder" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
+                     <padding>
+                        <Insets bottom="5" left="10" right="10" top="5" />
+                     </padding>
+                  </StackPane>
+
             <VBox fx:id="schedule" layoutX="560.0" layoutY="10.0" minWidth="240.0" prefWidth="240.0" styleClass="pane-with-border">
               <padding>
                 <Insets bottom="10" left="10" right="10" top="10" />
